work on MOW800 support

+++ sunray/ioboard.cpp
@@ -8,7 +8,7 @@
 // enable: true or false
 void ioI2cMux(uint8_t addr, uint8_t slave, bool enable){
   byte mask = (1 << slave);
-  Wire.requestFrom(addr, 1);
+  Wire.requestFrom(addr, (uint8_t)1);
   uint8_t state = Wire.read();   // get current control register state
   //CONSOLE.print("I2cMux control=");
   //CONSOLE.println(state, BIN);
@@ -29,7 +29,7 @@ void ioExpanderOut(uint8_t addr, uint8_t port, uint8_t pin, bool level){
   Wire.beginTransmission(addr); // PCA9555 address 
   Wire.write(6+port);    // configuration port    
   Wire.endTransmission();
-  Wire.requestFrom(addr, 1);  
+  Wire.requestFrom(addr, (uint8_t)1);  
   uint8_t state = Wire.read();   // get current configuration port
 
   Wire.beginTransmission(addr); // PCA9555 address 
@@ -40,7 +40,7 @@ void ioExpanderOut(uint8_t addr, uint8_t port, uint8_t pin, bool level){
   Wire.beginTransmission(addr); // PCA9555 address 
   Wire.write(2+port);    // output port    
   Wire.endTransmission();
-  Wire.requestFrom(addr, 1);  
+  Wire.requestFrom(addr, (uint8_t)1);  
   state = Wire.read();   // get current output port
 
   Wire.beginTransmission(addr); // PCA9555 address 
@@ -61,7 +61,7 @@ bool ioExpanderIn(uint8_t addr, uint8_t port, uint8_t pin){
   Wire.beginTransmission(addr); // PCA9555 address 
   Wire.write(6+port);    // configuration port    
   Wire.endTransmission();
-  Wire.requestFrom(addr, 1);  
+  Wire.requestFrom(addr, (uint8_t)1);  
   uint8_t state = Wire.read();   // get current configuration port
 
   Wire.beginTransmission(addr); // PCA9555 address 
@@ -72,7 +72,7 @@ bool ioExpanderIn(uint8_t addr, uint8_t port, uint8_t pin){
   Wire.beginTransmission(addr); // PCA9555 address 
   Wire.write(0+port);    // input port    
   Wire.endTransmission();
-  Wire.requestFrom(addr, 1);  
+  Wire.requestFrom(addr, (uint8_t)1);  
   state = Wire.read();   // get current output port
   return ((state & mask) != 0); 
 }
@@ -169,7 +169,7 @@ byte ioEepromReadByte( uint8_t addr, unsigned int eeaddress ) {
   Wire.write((int)(eeaddress >> 8)); // MSB
   Wire.write((int)(eeaddress & 0xFF)); // LSB
   Wire.endTransmission();
-  Wire.requestFrom(addr,1);
+  Wire.requestFrom(addr, (uint8_t)1);
   if (Wire.available()) rdata = Wire.read();
   return rdata;
 }
